+++ src/App.js
@@ -14,8 +14,7 @@ import * as Action from './model/state/actions';
 // Components
 import Login from './view/Login';
 import Profile from './view/Profile';
-import BoardContainer from './view/dashboard/board/BoardContainer.jsx';
-import TagContainer from './view/dashboard/board/TagContainer';
+import JobListings from './view/dashboard/containers/JobListings.jsx';
 import JobTracker from './view/dashboard/containers/JobTracker';
 
 
@@ -56,9 +55,8 @@ function App() {
         <Route exact path='/' component={Login}/>
         <Route path='/profile' component={Profile}/>
 
-        <Route path='/dashboard' component={BoardContainer} />
-        <Route path='/applied' component={JobTracker} />
-        <Route path='/tagged' component={TagContainer} />
+        <Route path='/dashboard' component={JobListings} />
+        {/* <Route path='/applied' component={JobTracker} /> */}
         <Route>Not Found</Route> {/* Catch all for non existing routes, will update with a 404 component that includes redirect to home page */}
       </Switch>
     </div>
