+++ src/store.js
@@ -1,6 +1,6 @@
 import { applyMiddleware, createStore, compose } from 'redux';
 import logger from 'redux-logger';
-import { reducer } from './model/state/reducer';
+import { reducer } from './model/state/jobs/reducer.js';
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
