--- src/view/dashboard/board/FeaturedPartners.jsx
@@ -1,19 +1,0 @@
-import React from 'react';
-
-const FeaturedPartners = () => {
-    return(
-        <div>
-            <h2>Featured Partners</h2>
-            <p>FP 1</p>
-            <p>FP 2</p>
-            <p>FP 3</p>
-            <p>FP 4</p>
-            <p>FP 5</p>
-            <p>FP 6</p>
-            <p>FP 7</p>
-            <p>FP 8</p>
-        </div>
-    )
-}
-
-export default FeaturedPartners;
