--- src/view/dashboard/board/SearchBar.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const SearchBar = () => {
-    return(
-        <div>
-            <input placeholder='⌘ K   Search Kondo Board' style={{width: '90%'}} />
-        </div>
-    )
-}
-
-export default SearchBar;
