--- src/view/dashboard/board/TagList.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const TagList = () => {
-    return(
-        <div></div>
-    )
-}
-
-export default TagList;
